Introduced BaseControllerHelperTrait
